ufo-galaxy-realization/requirements.txt
# UFO Galaxy L4 - 完整依赖列表
# Python 3.10+ 必需
# =================================

# ============================================================================
# 核心框架
# ============================================================================
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# ============================================================================
# 网络和通信
# ============================================================================
httpx>=0.25.0
aiohttp>=3.9.0
websockets>=11.0
requests>=2.31.0

# ============================================================================
# 数据库
# ============================================================================
sqlalchemy>=2.0.0
asyncpg>=0.29.0
aiosqlite>=0.19.0
redis>=4.6.0
qdrant-client>=1.6.0

# ============================================================================
# AI/LLM
# ============================================================================
openai>=1.0.0
anthropic>=0.7.0
google-generativeai>=0.3.0
tiktoken>=0.5.0

# ============================================================================
# Windows 客户端 UI
# ============================================================================
PyQt5>=5.15.0
pyqtwebengine>=5.15.0
pynput>=1.7.0
pywin32>=306; sys_platform == 'win32'

# ============================================================================
# 设备控制
# ============================================================================
pymavlink>=2.4.0
paho-mqtt>=1.6.0
pyserial>=3.5
bleak>=0.21.0
asyncssh>=2.14.0

# ============================================================================
# 媒体处理
# ============================================================================
pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
sounddevice>=0.4.0
edge-tts>=6.1.0

# ============================================================================
# 安全和加密
# ============================================================================
cryptography>=41.0.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0

# ============================================================================
# 工具和实用程序
# ============================================================================
pyyaml>=6.0
toml>=0.10.0
python-dateutil>=2.8.0
pytz>=2023.3

# ============================================================================
# 开发和测试
# ============================================================================
pytest>=7.4.0
pytest-asyncio>=0.21.0

# ============================================================================
# 打包
# ============================================================================
pyinstaller>=5.13.0

# ============================================================================
# 审计补充依赖 (Audit Added)
# ============================================================================
slack_sdk
qiskit
googlesearch-python
google-api-python-client
feedparser
aiortc
paddleocr
psutil
qiskit-aer
pytesseract
dbus-next
chromadb
qiskit-ibm-runtime
pyautogui
scikit-learn
av
pyperclip
beautifulsoup4
pygetwindow
pyserial-asyncio
